Added helper function Bob.extractToFolder with optional delete on exit

+++ com.dynamo.cr/com.dynamo.cr.bob/src/com/dynamo/bob/Bob.java
@@ -49,14 +49,12 @@ import org.apache.commons.cli.Options;
 import org.apache.commons.cli.ParseException;
 import org.apache.commons.cli.PosixParser;
 import org.apache.commons.io.FileUtils;
-import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.io.IOUtils;
 
 import com.dynamo.bob.archive.EngineVersion;
 import com.dynamo.bob.fs.DefaultFileSystem;
 import com.dynamo.bob.fs.IResource;
 import com.dynamo.bob.logging.Logger;
-import com.dynamo.bob.logging.LogFormatter;
 import com.dynamo.bob.logging.LogHelper;
 import com.dynamo.bob.util.LibraryUtil;
 import com.dynamo.bob.util.BobProjectProperties;
@@ -153,7 +151,7 @@ public class Bob {
         return rootFolder;
     }
 
-    public static void extract(final URL url, File toFolder) throws IOException {
+    public static void extractToFolder(final URL url, File toFolder, boolean deleteOnExit) throws IOException {
 
         ZipInputStream zipStream = new ZipInputStream(new BufferedInputStream(url.openStream()));
 
@@ -164,7 +162,8 @@ public class Bob {
                 if (!entry.isDirectory()) {
 
                     File dstFile = new File(toFolder, entry.getName());
-                    dstFile.deleteOnExit();
+                    if (deleteOnExit)
+                        dstFile.deleteOnExit();
                     dstFile.getParentFile().mkdirs();
 
                     OutputStream fileStream = null;
@@ -191,6 +190,10 @@ public class Bob {
         } finally {
             IOUtils.closeQuietly(zipStream);
         }
+    }
+
+    public static void extract(final URL url, File toFolder) throws IOException {
+        extractToFolder(url, toFolder, true);
     }
 
     public static String getPath(String path) {
